Move slice linop tests
